employee_pro/requirements.txt
certifi==2018.10.15
chardet==3.0.4

django-crispy-forms==1.7.2
idna==2.7
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23